Play around with a format I'd like to see.
I'm not sure what exactly to do with this, but my intent is to create an
alternative to XML/JSON. I'd like it to be a format that almost anyone
can edit, without being scared off by looking too much like "code".

Currently, it expresses two concepts (btw, this is very Razor
influenced).

Labeled metadata: All lines that start with @ represent a named element.

Hierarchy: Braces can be used for grouping.

And that's it.

So far, it feels really nice to read/write, with a lot of options
for how it _could_ work in the future. I'm just going to keep working
with it and see how I feel in a bit.